% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_trends.R
\name{run_mk}
\alias{run_mk}
\title{Run Modified Mann-Kendall Test}
\usage{
run_mk(x)
}
\arguments{
\item{x}{Numeric vector of data.}
}
\value{
A list containing Zc, p-value, Tau, and Sen's Slope.
}
\description{
Applies the Modified Mann-Kendall test with Sen's Slope estimator.
}
