% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_trends.R
\name{analyze_trends}
\alias{analyze_trends}
\title{Analyze Trends for All Pixels}
\usage{
analyze_trends(evi_data, start_year = NULL, end_year = NULL)
}
\arguments{
\item{start_year}{Optional integer. Start year for analysis.}

\item{end_year}{Optional integer. End year for analysis.}
}
\value{
An sf object with analysis results and geometry for each pixel.
}
\description{
Master function to apply BFAST, SSF, and MK tests to a dataset.
}
