% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_trends.R
\name{run_bfast}
\alias{run_bfast}
\title{Run BFAST Analysis}
\usage{
run_bfast(ts_data)
}
\arguments{
\item{ts_data}{A time series object (ts).}
}
\value{
The year of the first detected breakpoint, or NA if none found.
}
\description{
Applies BFAST Lite to detect breakpoints in the time series.
}
