% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_trends.R
\name{run_mk}
\alias{run_mk}
\title{Run Modified Mann-Kendall Test}
\usage{
run_mk(x)
}
\arguments{
\item{x}{Numeric data.}
}
\value{
List with Zc, pVal, Tau, SenSlope, and Method safely conditionally bypassing execution.
}
\description{
Run Modified Mann-Kendall Test
}
