% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_trends.R
\name{run_ssf_constrained}
\alias{run_ssf_constrained}
\title{Run ShapeSelectForest Analysis (Constrained)}
\usage{
run_ssf_constrained(y_val, x_pred, flat = TRUE, dec = TRUE, inc = TRUE)
}
\arguments{
\item{y_val}{Numeric response variable.}

\item{x_pred}{Numeric predictor variable.}

\item{flat}{Logical.}

\item{dec}{Logical.}

\item{inc}{Logical.}
}
\value{
Detected shape code array metric.
}
\description{
Run ShapeSelectForest Analysis (Constrained)
}
